tweet-service/src/main/java/com/gmail/mapper/BookmarkMapper.java
package com.gmail.mapper;

import com.gmail.dto.HeaderResponse;
import com.gmail.dto.response.tweet.TweetResponse;
import com.gmail.repository.projection.BookmarkProjection;
import com.gmail.repository.projection.TweetProjection;
import com.gmail.service.BookmarkService;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Component;

import java.util.ArrayList;
import java.util.List;

@Component
@RequiredArgsConstructor
public class BookmarkMapper {

    private final BasicMapper basicMapper;
    private final BookmarkService bookmarkService;

    public HeaderResponse<TweetResponse> getUserBookmarks(Pageable pageable) {
        Page<BookmarkProjection> bookmarks = bookmarkService.getUserBookmarks(pageable);
        List<TweetProjection> tweets = new ArrayList<>();
        bookmarks.getContent().forEach(bookmark -> tweets.add(bookmark.getTweet()));
        return basicMapper.getHeaderResponse(tweets, bookmarks.getTotalPages(), TweetResponse.class);
    }

    public Boolean processUserBookmarks(Long tweetId) {
        return bookmarkService.processUserBookmarks(tweetId);
    }

    public Boolean getIsTweetBookmarked(Long tweetId) {
        return bookmarkService.getIsTweetBookmarked(tweetId);
    }
}
